requirements.txt
fastapi[all]
openai
pandas
python-dotenv
thefuzz
python-Levenshtein
openpyxl
streamlit
requests
